backend/requirements.txt
Pillow==8.0.1
numpy==1.19.4